feat:refactor - register and login splitted
