+++ routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
-use App\infrastructure\Http\Controllers\Auth\AuthController;
+use App\infrastructure\Http\Controllers\Auth\OauthLoginAction;
+use App\infrastructure\Http\Controllers\Auth\RegisterOauthClientAction;
 use App\infrastructure\Http\Controllers\favorites\NewFavoriteGifAction;
 use App\infrastructure\Http\Controllers\Gifs\GetGifByIdAction;
 use App\infrastructure\Http\Controllers\Gifs\GetGifsBySpecificationAction;
@@ -17,20 +18,24 @@ use Illuminate\Support\Facades\Route;
 | be assigned to the "api" middleware group. Make something great!
 |
 */
-Route::get('health', [HealthCheckAction::class, 'execute']);
 
 
 Route::middleware('log.interactions')->group(function () {
+    Route::get('health', [HealthCheckAction::class, 'execute']);
+
+    Route::post('login', [OauthLoginAction::class, 'execute']);
+    Route::post('register', [RegisterOauthClientAction::class, 'execute']);
+
+
     Route::get('gifs/search', [GetGifsBySpecificationAction::class, 'execute']);
     Route::get('gifs/{id}', [GetGifByIdAction::class, 'execute']);
     Route::post('favorites', [NewFavoriteGifAction::class, 'execute']);
+//
+//    Route::middleware('auth:api')->group(function () {
+//    });
 });
 
-Route::post('login', [AuthController::class, 'login']);
-Route::post('register', [AuthController::class, 'register']);
-Route::middleware('auth:api')->get('user', function (Request $request) {
-    return $request->user();
-});
+
 
 Route::get('logs', [\App\infrastructure\Http\Controllers\LogController::class, 'showLogs']);
 //Route::post('register', [AuthController::class, 'register']);
